Štar version 3
in-range has gone: it was far, far too complicated for its own good,
and always buggy.  This is an incompatible change (but iterators have
never been stable).

the stepping iterators have optimizers, and are very fast.  There is a
new in-naturals iterator for simple dotime-like iteration which also
can be very fast.

There's a public star-syntax-error condition type.